client/requirements.txt
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0